+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Improved caching of vue component plugins to speed up loading
+- Introduced slot system so plugins can render arbitrary components on pages
+- Converted asset page to be a plugin
+- Refactored asset actions to be rendered by a page slot plugin
+
 ## [0.1.1] - 2022-05-27
 
 ### Added